[LUS] 0623 VM SDK Update - ResoureManager -> change to singleton - Rename: ExportedTextPNGSequenceData -> TextExportData   - add some fields - Change EditedData to own ResourceData directly - Changed auto_edit_helper to specify translate and scale values - Change text scale to be calculated from generatedVideo - Add Resolution Class - AutoEditedData -> delete map - StickerData -> add x, y

+++ lib/vm_sdk/types/resource.dart
@@ -132,8 +132,8 @@ class StickerData extends ResourceData {
   ResourceFileInfo? fileinfo;
   EStickerType type = EStickerType.object;
 
-  int x = 0;
-  int y = 0;
+  double x = 0;
+  double y = 0;
   double rotate = 0;
   double scale = 0;
